Merged from gioblu master

+++ examples/ARDUINO/Local/ThroughLoRa/HalfDuplex/Device2/Device2.ino
@@ -36,7 +36,7 @@ void receiver_function(uint8_t *payload, uint16_t length, const PJON_Packet_Info
 	}
 }
 
-void error_handler(uint8_t code, uint8_t data, void *custom_pointer) {
+void error_handler(uint8_t code, uint16_t data, void *custom_pointer) {
 	if (code == PJON_CONNECTION_LOST) {
 		Serial.print("Connection with device ID ");
 		Serial.print(bus.packets[data].content[0], DEC);
